refactor: 공용 훅 common 폴더로 이동
