+++ src/app/mypage/_components/UserDetails.tsx
@@ -1,5 +1,5 @@
 import Image from 'next/image';
-import useUser from '@hooks/useUser';
+import useUser from '@hooks/common/useUser';
 
 const UserDetails = () => {
   const { user, isPending } = useUser();
