+++ src/hooks/common/useDotIndicator.tsx
