+++ src/hooks/management/useRefuseNewMember.ts
@@ -3,7 +3,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { deleteMember } from '@queries/management/manageMembers';
 import { GroupsType, UsersType } from '@ts/supabaseTableRowTypes';
-import useUser from '@hooks/useUser';
+import useUser from '@hooks/common/useUser';
 
 interface UseRefuseNewMemberParams {
   groupId: GroupsType['id'];
